+++ src/App.jsx
@@ -1,6 +1,5 @@
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom'
 import './sass/main.scss'
-import Header from './components/header'
 import Section from './components/section'
 import Content from './components/content'
 import Hero from './components/hero'
@@ -10,10 +9,6 @@ import Footer from './components/footer'
 const App = () => {
   return (
     <>
-      <Header 
-        headerTitle="4 Petes Sake"
-      >
-      </Header>
       <Section>
         <Hero>this is the hero section</Hero>
       </Section>
