+++ src/stories/Header.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Header } from './Header';
+import { Header } from '../components/header'
 
 export default {
   title: 'Example/Header',
