+++ src/stories/Page.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Header } from './Header';
+import { Header } from '../components/header'
 import './page.css';
 
 export const Page = () => {
